# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "server_for_EnergoGAZ",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Supporting documentation for working with the server part of the EnergoGAZ company's accounting system."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/work/allEmployees": {
      "get": {
        "tags": [
          "Operational Work"
        ],
        "summary": "List of employees",
        "operationId": "allEmployees",
        "description": "List of employees",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All employeess",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "All employeess"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "_id",
                          "name",
                          "jobTitle",
                          "placeOfWork"
                        ],
                        "properties": {
                          "_id": {
                            "description": "employeess id",
                            "type": "string",
                            "example": "6a22c356b0a3c31e140331a0"
                          },
                          "name": {
                            "description": "employeess name",
                            "type": "string",
                            "example": "Колобков"
                          },
                          "jobTitle": {
                            "description": "job title",
                            "type": "string",
                            "example": "менеджер"
                          },
                          "placeOfWork": {
                            "description": "place of work",
                            "type": "string",
                            "example": "офис"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
...
    "/api/work/addingAgzp": {
      "post": {
        "tags": [
          "Operational Work"
        ],
        "summary": "Adding a new gas station",
        "operationId": "addingAgzp",
        "description": "Adding a new gas station",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "gasStationNumber",
                  "typeOfFuel",
                  "address",
                  "phoneNumber",
                  "listOfEmployees",
                  "fuelTanks",
                  "dispenserName",
                  "commissioningDate",
                  "lastMaintenance"
                ],
                "properties": {
                  "gasStationNumber": {
                    "description": "gas station number",
                    "type": "string",
                    "example": "agzp_22"
                  },
                  "typeOfFuel": {
                    "description": "type of fuel",
                    "type": "array",
                    "example": [
                      {
                        "view": "gaz"
                      }
                    ]
                  },
                  "address": {
                    "description": "address agzp",
                    "type": "string",
                    "example": "Черкаська обл. м. Хлебное, вул. Ватутена, 28"
                  },
                  "phoneNumber": {
                    "description": "phone number",
                    "type": "string",
                    "example": "123-123-45-67"
                  },
                  "listOfEmployees": {
                    "description": "list of employees",
                    "type": "array",
                    "example": [
                      "Козлов",
                      "Петров",
                      "Сидоров"
                    ]
                  },
                  "fuelTanks": {
                    "description": "fuel tanks",
                    "type": "array",
                    "example": [
                      {
                        "containerNumber": 1,
                        "volume": 9960
                      },
                      {
                        "containerNumber": 2,
                        "volume": 10050
                      }
                    ]
                  },
                  "dispenserName": {
                    "description": "dispenser name",
                    "type": "string",
                    "example": "shelf"
                  },
                  "commissioningDate": {
                    "description": "commissioning date",
                    "type": "string",
                    "example": "01.01.2014"
                  },
                  "lastMaintenance": {
                    "description": "last maintenance",
                    "type": "string",
                    "example": "12.02.2022"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a new agzp",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "AGZP added"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "gasStationNumber",
                        "typeOfFuel",
                        "address",
                        "phoneNumber",
                        "listOfEmployees",
                        "fuelTanks",
                        "dispenserName",
                        "commissioningDate",
                        "lastMaintenance",
                        "_id"
                      ],
                      "properties": {
                        "gasStationNumber": {
                          "description": "gas station number",
                          "type": "string",
                          "example": "agzp_22"
                        },
                        "typeOfFuel": {
                          "description": "type of fuel",
                          "type": "array",
                          "example": [
                            {
                              "view": "gaz"
                            }
                          ]
                        },
                        "address": {
                          "description": "address agzp",
                          "type": "string",
                          "example": "Черкаська обл. м. Хлебное, вул. Ватутена, 28"
                        },
                        "phoneNumber": {
                          "description": "phone number",
                          "type": "string",
                          "example": "123-123-45-67"
                        },
                        "listOfEmployees": {
                          "description": "list of employees",
                          "type": "array",
                          "example": [
                            "Козлов",
                            "Петров",
                            "Сидоров"
                          ]
                        },
                        "fuelTanks": {
                          "description": "fuel tanks",
                          "type": "array",
                          "example": [
                            {
                              "containerNumber": 1,
                              "volume": 9960
                            },
                            {
                              "containerNumber": 2,
                              "volume": 10050
                            }
                          ]
                        },
                        "dispenserName": {
                          "description": "dispenser name",
                          "type": "string",
                          "example": "shelf"
                        },
                        "commissioningDate": {
                          "description": "commissioning date",
                          "type": "string",
                          "example": "01.01.2014"
                        },
                        "lastMaintenance": {
                          "description": "last maintenance",
                          "type": "string",
                          "example": "12.02.2022"
                        },
                        "_id": {
                          "description": "id agzp",
                          "type": "string",
                          "example": "6a3bda1ad130661459b25c6c"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
    "/api/work/allAgzp": {
      "get": {
        "tags": [
          "Operational Work"
        ],
        "summary": "List of all gas stations",
        "operationId": "allAgzp",
        "description": "List of all gas stations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All AGZP",
            "content": {
              "aplication-json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "All AGZP"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "gasStationNumber",
                          "typeOfFuel",
                          "address",
                          "phoneNumber",
                          "listOfEmployees",
                          "fuelTanks",
                          "dispenserName",
                          "commissioningDate",
                          "lastMaintenance",
                          "_id",
                          "userId",
                          "userName"
                        ],
                        "properties": {
                          "gasStationNumber": {
                            "description": "gas station number",
                            "type": "string",
                            "example": "agzp_22"
                          },
                          "typeOfFuel": {
                            "description": "type of fuel",
                            "type": "array",
                            "example": [
                              {
                                "view": "gaz"
                              }
                            ]
                          },
                          "address": {
                            "description": "address agzp",
                            "type": "string",
                            "example": "Черкаська обл. м. Хлебное, вул. Ватутена, 28"
                          },
                          "phoneNumber": {
                            "description": "phone number",
                            "type": "string",
                            "example": "123-123-45-67"
                          },
                          "listOfEmployees": {
                            "description": "list of employees",
                            "type": "array",
                            "example": [
                              "Козлов",
                              "Петров",
                              "Сидоров"
                            ]
                          },
                          "fuelTanks": {
                            "description": "fuel tanks",
                            "type": "array",
                            "example": [
                              {
                                "containerNumber": 1,
                                "volume": 9960
                              },
                              {
                                "containerNumber": 2,
                                "volume": 10050
                              }
                            ]
                          },
                          "dispenserName": {
                            "description": "dispenser name",
                            "type": "string",
                            "example": "shelf"
                          },
                          "commissioningDate": {
                            "description": "commissioning date",
                            "type": "string",
                            "example": "01.01.2014"
                          },
                          "lastMaintenance": {
                            "description": "last maintenance",
                            "type": "string",
                            "example": "12.02.2022"
                          },
                          "_id": {
                            "description": "employeess id",
                            "type": "string",
                            "example": "6a3bda1ad130661459b25c6c"
                          },
                          "userId": {
                            "description": "Company employee ID",
                            "type": "string",
                            "example": "6a22c356b0a3c31e140331a0"
                          },
                          "userName": {
                            "description": "Name of the company employee",
                            "type": "string",
                            "example": "Иванов"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
...
    "/api/work/updateDailyReportAgzp/{dailyReportId}": {
      "patch": {
        "tags": [
          "Operational Work"
        ],
        "summary": "Daily report update",
        "operationId": "updateDailyReportAgzp",
        "description": "Daily report update",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "dailyReportId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6a3fdc5619aaaac34baa0513"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numberAgzp": {
                    "description": "gas station number",
                    "type": "string",
                    "example": "agzp_1"
                  },
                  "date": {
                    "description": "What period is the report for",
                    "type": "string",
                    "example": "2026-02-25"
                  },
                  "operatorName": {
                    "description": "Operators name",
                    "type": "string",
                    "example": "Иванов И.И."
                  },
                  "typeOfFuel": {
                    "description": "type of fuel",
                    "type": "string",
                    "enum": [
                      "gas",
                      "gasoline_A_95",
                      "gasoline_A_92",
                      "diesel"
                    ]
                  },
                  "fuelArrival": {
                    "description": "arrival of fuel at the gas station",
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "containerNumber": {
                          "type": "number",
                          "description": "Tank number"
                        },
                        "volume": {
                          "type": "number",
                          "description": "Drained fuel"
                        }
                      },
                      "required": [
                        "containerNumber",
                        "volume"
                      ]
                    },
                    "example": [
                      {
                        "containerNumber": 1,
                        "volume": 0
                      },
                      {
                        "containerNumber": 2,
                        "volume": 10000
                      }
                    ]
                  },
                  "noteForFuelArrival": {
                    "description": "Note for fuel arrival",
                    "type": "string",
                    "example": "в 1 емкость слито 5000 л., во 2 емкость слито 4000 л."
                  },
                  "initialCounterValue1": {
                    "description": "Initial counter value 1",
                    "type": "number",
                    "example": 2567874.23
                  },
                  "finalMeterReading1": {
                    "description": "Final meter reading 1",
                    "type": "number",
                    "example": 2569874.23
                  },
                  "initialCounterValue2": {
                    "description": "Initial counter value 2",
                    "type": "number",
                    "example": 0
                  },
                  "finalMeterReading2": {
                    "description": "Final meter reading 2",
                    "type": "number",
                    "example": 0
                  },
                  "fuelResidues": {
                    "description": "Remaining fuel at the gas station",
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "containerNumber": {
                          "type": "number",
                          "description": "Tank number"
                        },
                        "remainder": {
                          "type": "number",
                          "description": "Remaining fuel"
                        }
                      },
                      "required": [
                        "containerNumber",
                        "remainder"
                      ]
                    },
                    "example": [
                      {
                        "containerNumber": 1,
                        "remainder": 85
                      },
                      {
                        "containerNumber": 2,
                        "remainder": 65
                      }
                    ]
                  },
                  "noteOnFuelRemaining": {
                    "description": "Note on fuel remaining",
                    "type": "string",
                    "example": ""
                  },
                  "price": {
                    "description": "Price of fuel at the gas stationl",
                    "type": "number",
                    "example": 28.5
                  },
                  "cashRegisterTotal": {
                    "description": "Cash register amount per shift",
                    "type": "number",
                    "example": 3654.5
                  },
                  "byTerminal": {
                    "description": "Amount per terminal per shift",
                    "type": "number",
                    "example": 2521.25
                  },
                  "litersByCashRegister": {
                    "description": "Liters by cash register and terminal",
                    "type": "number",
                    "example": 256.5
                  },
                  "f2Total": {
                    "description": "Sum over f2",
                    "type": "number",
                    "example": 12035.25
                  },
                  "litersByF2": {
                    "description": "Liters by f2",
                    "type": "number",
                    "example": 505.5
                  },
                  "accordingToContract": {
                    "description": "Refueling under contract",
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "nameCompani": {
                          "type": "string",
                          "description": "Company name"
                        },
                        "liters": {
                          "type": "number",
                          "description": "Liters filled"
                        }
                      }
                    },
                    "example": [
                      {
                        "nameCompani": "МХП",
                        "liters": 250
                      },
                      {
                        "nameCompani": "Коло",
                        "liters": 50
                      }
                    ]
                  },
                  "fuelForTheOffice": {
                    "description": "Fuel for office needs",
                    "type": "number",
                    "example": 50
                  },
                  "noteForFuelForTheOffice": {
                    "description": "Office fuel note",
                    "type": "string",
                    "example": "30 - управляющий АГЗП, 20 - главный инженер"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/200updateDailyReport"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "NotFoundError",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "No such report found."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/work/deleteCompanie/{companieId}": {
      "delete": {
        "tags": [
          "Operational Work"
        ],
        "summary": "Deleting a company by ID",
        "operationId": "deleteCompanie",
        "description": "Deleting a company by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "companieId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6a42592e1332b645bbbf0568"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The company has been deleted."
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "NotFoundError",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "No such company was found."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "401": {
        "description": "UnauthorizedError",
        "content": {
          "application/json": {
            "schema": {
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Session not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "ForbiddenError",
        "content": {
          "application/json": {
            "schema": {
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 403
                },
                "message": {
                  "type": "string",
                  "example": "ForbiddenError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access denied"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}